+++ src/pages/main-screen/main-screen.tsx
@@ -1,18 +1,12 @@
 // noinspection JSDeprecatedSymbols
 
-import {useId} from 'react';
-import PlaceCard from '../../components/blocks/place-card/place-card.tsx';
-import {ReactNode} from 'react';
-import {OFFERS_SHOW_COUNT} from '../../data/magic-numbers.ts';
+import CityOffers from '../../components/blocks/city-offers/city-offers.tsx';
 import Header from '../../components/layout/header/header.tsx';
+import Tabs from '../../components/blocks/tabs/tabs.tsx';
+import {OFFERS_SHOW_COUNT} from '../../data/magic-numbers.ts';
 
 
-type MainScreenProps = {
-  offersCount: number;
-}
-
-export default function MainScreen ({offersCount}: MainScreenProps): JSX.Element{
-  const uniqueKey = useId();
+export default function MainScreen (): JSX.Element{
   return (
 
     <div className="page page--gray page--main">
@@ -22,94 +16,12 @@ export default function MainScreen ({offersCount}: MainScreenProps): JSX.Element
         <h1 className="visually-hidden">Cities</h1>
         <div className="tabs">
           <section className="locations container">
-            <ul className="locations__list tabs__list">
-              <li className="locations__item">
-                <a className="locations__item-link tabs__item"
-                  href="#"
-                >
-                  <span>Paris</span>
-                </a>
-              </li>
-              <li className="locations__item">
-                <a className="locations__item-link tabs__item"
-                  href="#"
-                >
-                  <span>Cologne</span>
-                </a>
-              </li>
-              <li className="locations__item">
-                <a className="locations__item-link tabs__item"
-                  href="#"
-                >
-                  <span>Brussels</span>
-                </a>
-              </li>
-              <li className="locations__item">
-                <a className="locations__item-link tabs__item tabs__item--active">
-                  <span>Amsterdam</span>
-                </a>
-              </li>
-              <li className="locations__item">
-                <a className="locations__item-link tabs__item"
-                  href="#"
-                >
-                  <span>Hamburg</span>
-                </a>
-              </li>
-              <li className="locations__item">
-                <a className="locations__item-link tabs__item"
-                  href="#"
-                >
-                  <span>Dusseldorf</span>
-                </a>
-              </li>
-            </ul>
+            <Tabs/>
           </section>
         </div>
         <div className="cities">
           <div className="cities__places-container container">
-            <section className="cities__places places">
-              <h2 className="visually-hidden">Places</h2>
-              <b className="places__found">{offersCount} places to stay in Amsterdam</b>
-              <form className="places__sorting"
-                action="#"
-                method="get"
-              >
-                <span className="places__sorting-caption">Sort by</span>
-                <span className="places__sorting-type"
-                  tabIndex={0}
-                >
-                  Popular
-                  <svg className="places__sorting-arrow"
-                    width="7"
-                    height="4"
-                  >
-                    <use xlinkHref="#icon-arrow-select"></use>
-                  </svg>
-                </span>
-                <ul className="places__options places__options--custom places__options--opened">
-                  <li className="places__option places__option--active"
-                    tabIndex={0}
-                  >Popular
-                  </li>
-                  <li className="places__option"
-                    tabIndex={0}
-                  >Price: low to high
-                  </li>
-                  <li className="places__option"
-                    tabIndex={0}
-                  >Price: high to low
-                  </li>
-                  <li className="places__option"
-                    tabIndex={0}
-                  >Top rated first
-                  </li>
-                </ul>
-              </form>
-              <div className="cities__places-list places__list tabs__content">
-                {Array.from({length: OFFERS_SHOW_COUNT}).map(():ReactNode => <PlaceCard key={uniqueKey}/>)}
-              </div>
-            </section>
+            <CityOffers offersCount={OFFERS_SHOW_COUNT}/>
             <div className="cities__right-section">
               <section className="cities__map map"></section>
             </div>
